refactor(chatbot): extract core chatbot logic into separate module
- Moved core chatbot logic out of handler function, into separate functions, into their own file
